Added npm / yarn installation instructions.
This commit attempts to clarify the commands used for
the bootstrapping and the client install via npm or yarn.
In addition, it explicitly states the version of npm and
yarn that support the init/create commands required for
bootstrapping.  Debian based distros often come with ancient
versions of npm that support init but not in a way
that will work with this feature.